+++ server/src/main/java/org/opensearch/cluster/node/DiscoveryNodeRole.java
@@ -298,13 +298,13 @@ public abstract class DiscoveryNodeRole implements Comparable<DiscoveryNodeRole>
     };
 
     /**
-     * Represents the role for a search node, which is dedicated to provide search capability.
+     * Represents the role for a warm node, which is dedicated to hold warm indices.
      */
-    public static final DiscoveryNodeRole SEARCH_ROLE = new DiscoveryNodeRole("search", "s", true) {
+    public static final DiscoveryNodeRole WARM_ROLE = new DiscoveryNodeRole("warm", "w", true) {
 
         @Override
         public Setting<Boolean> legacySetting() {
-            // search role is added in 2.4 so doesn't need to configure legacy setting
+            // warm role is added in 2.4 so doesn't need to configure legacy setting
             return null;
         }
 
@@ -314,7 +314,7 @@ public abstract class DiscoveryNodeRole implements Comparable<DiscoveryNodeRole>
      * The built-in node roles.
      */
     public static SortedSet<DiscoveryNodeRole> BUILT_IN_ROLES = Collections.unmodifiableSortedSet(
-        new TreeSet<>(Arrays.asList(DATA_ROLE, INGEST_ROLE, CLUSTER_MANAGER_ROLE, REMOTE_CLUSTER_CLIENT_ROLE, SEARCH_ROLE))
+        new TreeSet<>(Arrays.asList(DATA_ROLE, INGEST_ROLE, CLUSTER_MANAGER_ROLE, REMOTE_CLUSTER_CLIENT_ROLE, WARM_ROLE))
     );
 
     /**